+++ ICD.Connect.Panels.Crestron/Controls/TouchScreens/IThreeSeriesTouchScreenControl.cs
@@ -1,4 +1,4 @@
-﻿using ICD.Connect.Panels.Controls;
+﻿using ICD.Connect.Panels.Controls.Panels;
 
 namespace ICD.Connect.Panels.Crestron.Controls.TouchScreens
 {