+++ ICD.Connect.Panels.CrestronPro/TriListAdapters/Controls/HardButtons/TswX60HardButtonBacklightControl.cs
@@ -1,12 +1,12 @@
 ﻿using ICD.Common.Utils;
 using ICD.Common.Utils.Services.Logging;
 using ICD.Connect.Devices.EventArguments;
+using ICD.Connect.Panels.Controls.HardButtons;
 #if SIMPLSHARP
 using System;
 using System.Collections.Generic;
 using Crestron.SimplSharpPro.UI;
 using ICD.Connect.API.Commands;
-using ICD.Connect.Panels.Controls;
 using ICD.Connect.Panels.CrestronPro.TriListAdapters.X60;
 
 namespace ICD.Connect.Panels.CrestronPro.TriListAdapters.Controls.HardButtons