+++ ICD.Connect.Panels/Controls/Sigs/ISigControl.cs
@@ -1,6 +1,6 @@
 ﻿using ICD.Connect.Devices.Controls;
 
-namespace ICD.Connect.Panels.Controls
+namespace ICD.Connect.Panels.Controls.Sigs
 {
 	public interface ISigControl : IDeviceControl, ISigInputOutput
 	{